backend/requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
flask-cors
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
